+++ src/holiday/holiday.test.ts
@@ -1,11 +1,11 @@
 import type { BasicJewishDate } from "jewish-date";
 import { describe, expect, it } from "vitest";
-import { isHoliday } from "../../src";
-import type { Holiday } from "../../src/interfaces";
+import { isDateInHolidayList } from ".";
+import type { Holiday } from "../interfaces";
 
 describe("Holiday validation", () => {
   it("should return false when given null as BasicJewishDate and holidayList is empty", () => {
-    const result = isHoliday(null as unknown as BasicJewishDate, []);
+    const result = isDateInHolidayList(null as unknown as BasicJewishDate, []);
     expect(result).toBe(false);
   });
 
@@ -19,7 +19,7 @@ describe("Holiday validation", () => {
       { day: 1, monthName: "Tishri", name: "Rosh Hashanah" },
     ];
 
-    const result = isHoliday(jewishDate, holidayList);
+    const result = isDateInHolidayList(jewishDate, holidayList);
     expect(result).toBe(true);
   });
 });